+++ json_serializable/test/default_value/default_value_interface.dart
@@ -6,8 +6,6 @@
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
 
-// ignore_for_file: annotate_overrides
-
 abstract class DefaultValue {
   bool fieldBool;
   String fieldString;