+++ json_serializable/test/generic_files/generic_class.g.dart
@@ -30,17 +30,17 @@ GenericClass<T, S> _$GenericClassFromJson<T extends num, S>(
         : _dataFromJson(json['fieldS'] as Map<String, dynamic>);
 }
 
-abstract class _$GenericClassSerializerMixin<T extends num, S> {
-  Object get fieldObject;
-  dynamic get fieldDynamic;
-  int get fieldInt;
-  T get fieldT;
-  S get fieldS;
-  Map<String, dynamic> toJson() => <String, dynamic>{
-        'fieldObject': fieldObject == null ? null : _dataToJson(fieldObject),
-        'fieldDynamic': fieldDynamic == null ? null : _dataToJson(fieldDynamic),
-        'fieldInt': fieldInt == null ? null : _dataToJson(fieldInt),
-        'fieldT': fieldT == null ? null : _dataToJson(fieldT),
-        'fieldS': fieldS == null ? null : _dataToJson(fieldS)
-      };
-}
+Map<String, dynamic> _$GenericClassToJson<T extends num, S>(
+        GenericClass<T, S> instance) =>
+    <String, dynamic>{
+      'fieldObject': instance.fieldObject == null
+          ? null
+          : _dataToJson(instance.fieldObject),
+      'fieldDynamic': instance.fieldDynamic == null
+          ? null
+          : _dataToJson(instance.fieldDynamic),
+      'fieldInt':
+          instance.fieldInt == null ? null : _dataToJson(instance.fieldInt),
+      'fieldT': instance.fieldT == null ? null : _dataToJson(instance.fieldT),
+      'fieldS': instance.fieldS == null ? null : _dataToJson(instance.fieldS)
+    };