+++ json_serializable/tool/build.dart
@@ -13,11 +13,13 @@ import 'package:json_serializable/src/json_part_builder.dart' as jpb;
 
 import 'builder.dart';
 
-Builder _jsonPartBuilder(
-    {bool useWrappers: false,
-    bool anyMap: false,
-    bool checked: false,
-    bool format}) {
+Builder _jsonPartBuilder({
+  bool useWrappers,
+  bool anyMap,
+  bool checked,
+  bool format,
+  bool generateToJsonFunction,
+}) {
   format ??= true;
   String Function(String code) formatOutput;
 
@@ -30,7 +32,8 @@ Builder _jsonPartBuilder(
       formatOutput: formatOutput,
       useWrappers: useWrappers,
       anyMap: anyMap,
-      checked: checked);
+      checked: checked,
+      generateToJsonFunction: generateToJsonFunction);
 }
 
 final List<BuilderApplication> builders = [
@@ -56,11 +59,16 @@ final List<BuilderApplication> builders = [
       generateFor: const InputSet(
         include: const [
           'example/example.dart',
-          'test/default_value/default_value.dart',
-          'test/generic_files/generic_class.dart',
           'test/test_files/json_literal.dart',
           'test/test_files/json_test_example.dart',
           'test/test_files/json_test_example.non_nullable.dart'
+        ],
+      )),
+  applyToRoot(_jsonPartBuilder(generateToJsonFunction: true),
+      generateFor: const InputSet(
+        include: const [
+          'test/generic_files/generic_class.dart',
+          'test/default_value/default_value.dart',
         ],
       )),
   applyToRoot(_jsonPartBuilder(anyMap: true),
@@ -75,16 +83,28 @@ final List<BuilderApplication> builders = [
   applyToRoot(_jsonPartBuilder(checked: true, anyMap: true),
       generateFor: const InputSet(
         include: const [
-          'test/default_value/default_value.checked.dart',
           'test/kitchen_sink/kitchen_sink.non_nullable.checked.dart',
           'test/yaml/build_config.dart',
+        ],
+      )),
+  applyToRoot(
+      _jsonPartBuilder(
+          checked: true, anyMap: true, generateToJsonFunction: true),
+      generateFor: const InputSet(
+        include: const [
+          'test/default_value/default_value.checked.dart',
         ],
       )),
   applyToRoot(_jsonPartBuilder(useWrappers: true),
       generateFor: const InputSet(
         include: const [
-          'test/generic_files/generic_class*wrapped.dart',
           'test/test_files/json_test_example*wrapped.dart',
+        ],
+      )),
+  applyToRoot(_jsonPartBuilder(useWrappers: true, generateToJsonFunction: true),
+      generateFor: const InputSet(
+        include: const [
+          'test/generic_files/generic_class*wrapped.dart',
         ],
       )),
   applyToRoot(_jsonPartBuilder(useWrappers: true, anyMap: true),